+++ packages/ui/src/react-native-web-augment.ts
@@ -3,6 +3,7 @@ import 'react-native';
 declare module 'react-native' {
   interface ViewProps {
     className?: string;
+    dataSet?: Record<string, string>;
   }
 
   interface TextProps {
@@ -15,14 +16,17 @@ declare module 'react-native' {
 
   interface PressableProps {
     className?: string;
+    dataSet?: Record<string, string>;
   }
 
   interface TouchableOpacityProps {
     className?: string;
+    dataSet?: Record<string, string>;
   }
 
   interface ScrollViewProps {
     className?: string;
+    dataSet?: Record<string, string>;
   }
 
   interface ImageBackgroundProps {